No more scripts. Added multi-line support for replace.
