+++ lib/fs-driver.js
@@ -241,9 +241,9 @@ FsDriver.prototype.grep = function (text, cb) {
  *   completes.
  */
 FsDriver.prototype.sed = function (search, replace, name, cb) {
-  var pattern = [
-    's', FsDriver.escape(search), FsDriver.escape(replace), 'g'
-  ].join('/');
+  var searchEsc = FsDriver.escape(search).replace(/\n/g, '\\n');
+  var replaceEsc = FsDriver.escape(replace).replace(/\n/g, '\\n');
+  var pattern = ['s', searchEsc, replaceEsc, 'g'].join('/');
 
   var command = [
     'sed -i.last',
