+++ lib/transformer.js
@@ -4,10 +4,10 @@ var async = require('async');
 var isString = require('101/is-string');
 var FsDriver = require('./fs-driver');
 var Warning = require('./warning');
-var ScriptGenerator = require('./script-generator');
 var exists = require('101/exists');
 var isEmpty = require('101/is-empty');
 var debug = require('debug');
+var fs = require('fs');
 
 var fullDiffDebug = debug('fs-transform:full-diff');
 
@@ -45,7 +45,6 @@ function Transformer(root, rules) {
   this.results = [];
   this.nameChanges = [];
   this._globalExcludes = [];
-  this.script = new ScriptGenerator();
 
   this.setAction('copy', this.copy);
   this.setAction('rename', this.rename);
@@ -115,14 +114,6 @@ Transformer.prototype.setFileDiff = function (filename, diff) {
   // absolute paths to the root directory.
   var relativePath = this.driver.stripAbsolutePaths(filename);
   this.currentResult.diffs[relativePath] = diff;
-};
-
-/**
- * Generates a shell script from this transfromer's rules.
- * @return {string} A shell-script that executes the rules.
- */
-Transformer.prototype.getScript = function () {
-  return this.script.generate();
 };
 
 /**
@@ -353,7 +344,6 @@ Transformer.prototype.copy = function (rule, cb) {
   this.driver.copy(rule.source, rule.dest, function (err, output) {
     if (err) { return cb(err); }
     self.addNameChange(rule.source, rule.dest);
-    self.script.addRule(rule);
     cb();
   });
 };
@@ -371,7 +361,6 @@ Transformer.prototype.rename = function (rule, cb) {
   this.driver.move(rule.source, rule.dest, function (err, output) {
     if (err) { return cb(err); }
     self.addNameChange(rule.source, rule.dest);
-    self.script.addRule(rule);
     cb();
   });
 };
@@ -413,7 +402,6 @@ Transformer.prototype.replace = function (rule, cb) {
     // 1.1 If no files were returned by grep, issue a warning and return
     if (files.length === 0) {
       self.addWarning(rule, 'Search did not return any results.');
-      self.script.addRule(rule);
       return cb();
     }
 
@@ -422,6 +410,16 @@ Transformer.prototype.replace = function (rule, cb) {
     // 2.0 Filter out global excludes and .git files
     files = files.filter(function (file) {
       return !~self._globalExcludes.indexOf(file) && !~file.indexOf('.git/');
+    });
+
+    // 2.0.1 Filter out files that do not exactly match pattern (for multiline)
+    files = files.filter(function (file) {
+      // TODO This could possibly use a lot of memory and is pretty slow.
+      //      Going to need a better way if this breaks. Happens because the
+      //      buffer from readFileSync is being toString-ed.
+      var path = self.driver.absolutePath(file);
+      var contents = fs.readFileSync(path).toString();
+      return contents.indexOf(rule.search) !== -1;
     });
 
     // 2.1 Keep track of when excludes are used to filter results
@@ -467,14 +465,20 @@ Transformer.prototype.replace = function (rule, cb) {
       },
 
       // 3.2 Make replacements by using sed
-      function sedFiles(cb) {
-        async.each(files, function (file, sedCallback) {
-          self.driver.sed(
-            search,
-            replace,
-            file,
-            sedCallback
-          );
+      function makeReplacements(cb) {
+        async.each(files, function (file, replaceCallback) {
+          var path = self.driver.absolutePath(file);
+          fs.readFile(file, function (err, data) {
+            if (err) { return replaceCallback(err); }
+            // TODO This could possibly use a lot of memory and is pretty slow.
+            //      Going to need a better way if this breaks. Happens because
+            //      the buffer from readFileSync is being toString-ed.
+            var contents = data.toString();
+            while (contents.indexOf(rule.search) !== -1) {
+              contents = contents.replace(rule.search, rule.replace);
+            }
+            fs.writeFile(path, contents, replaceCallback);
+          });
         }, cb);
       },
 
@@ -499,12 +503,9 @@ Transformer.prototype.replace = function (rule, cb) {
 
       // 3.4 Cleanup temporary files
       function cleanup(cb) {
-        self.script.addRule(rule);
         async.each(files, function (name, removeCallback) {
           var copyName = name + Transformer.ORIGINAL_POSTFIX;
-          var dotLastName = name + '.last';
-          var names = [copyName, dotLastName].join(' ');
-          self.driver.remove(names, removeCallback);
+          self.driver.remove(copyName, removeCallback);
         }, cb);
       }
     ], cb);
@@ -536,6 +537,5 @@ Transformer.prototype.exclude = function (rule, cb) {
     }
   });
 
-  self.script.addRule(rule);
   cb();
 };
