+++ test/fixtures/root/A
@@ -4,4 +4,8 @@ Example
 Exampel
 Interesting
 
+  Multiline
+EXAMPLE
+Drool...
+
 /some/path/foo
