+++ test/fixtures/root/B
@@ -5,4 +5,8 @@ File B is good
 Interesting
 Exampel
 
+Multiline
+EXAMPLE
+Super Cool
+
 Mew
