+++ test/unit/transformer/constructor.js
@@ -10,8 +10,8 @@ var after = lab.after;
 var afterEach = lab.afterEach;
 var Code = require('code');
 var expect = Code.expect;
-var Transformer = require('../../lib/transformer');
-var ScriptGenerator = require('../../lib/script-generator');
+
+var Transformer = require('../../../lib/transformer');
 
 describe('Transformer', function() {
   describe('constructor', function() {
@@ -51,12 +51,6 @@ describe('Transformer', function() {
     it('should keep a list of results', function(done) {
       var transformer = new Transformer('/etc', []);
       expect(transformer.results).to.be.an.array();
-      done();
-    });
-
-    it('should have a shell script generator', function(done) {
-      var transformer = new Transformer('/etc', []);
-      expect(transformer.script).instanceof(ScriptGenerator);
       done();
     });
